requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.14.0
fonttools==4.53.0
geojson==3.1.0
geojsoncontour==0.4.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.2.3
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
sentry-sdk==2.19.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
tenacity==9.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.5.20
uvicorn==0.34.0
virtualenv==20.26.2
watchfiles==1.0.4
websockets==14.1
xarray==2024.5.0
